# frontend/docs/swagger/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "ApiResponseCommentIdResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CommentIdResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
      "ApiResponseCommentPageResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CommentPageResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "ApiResponseDetailResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DetailResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "ApiResponseLikeToggleResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LikeToggleResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "ApiResponsePostDetailResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PostDetailResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
      "ApiResponsePostIdResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PostIdResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
      "ApiResponsePostPageResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PostPageResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "ApiResponseUnit": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "AuthorResponse": {
        "properties": {
          "memberId": {
            "format": "uuid",
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "profileImageUrl": {
            "type": "string"
          }
        },
        "required": [
          "memberId"
        ],
        "type": "object"
      },
...
      "CommentIdResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
      "CommentPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CommentResponse"
            },
            "type": "array"
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "CommentResponse": {
        "properties": {
          "author": {
            "$ref": "#/components/schemas/AuthorResponse"
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "parentCommentId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "author",
          "body",
          "createdAt",
          "deleted",
          "id"
        ],
        "type": "object"
      },
...
      "ConditionResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "text"
        ],
        "type": "object"
      },
...
      "CreateCommentRequest": {
        "properties": {
          "body": {
            "type": "string"
          },
          "mediaId": {
            "format": "uuid",
            "type": "string"
          },
          "parentCommentId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "type": "object"
      },
...
      "CurrentResponse": {
        "properties": {
          "condition": {
            "$ref": "#/components/schemas/ConditionResponse"
          },
          "feelsLikeTemperature": {
            "format": "int32",
            "type": "integer"
          },
          "humidity": {
            "format": "int32",
            "type": "integer"
          },
          "maxTemperature": {
            "format": "int32",
            "type": "integer"
          },
          "minTemperature": {
            "format": "int32",
            "type": "integer"
          },
          "precipitationProbability": {
            "format": "int32",
            "type": "integer"
          },
          "temperature": {
            "format": "int32",
            "type": "integer"
          },
          "uvIndex": {
            "format": "int32",
            "type": "integer"
          },
          "windSpeed": {
            "format": "double",
            "type": "number"
          }
        },
        "required": [
          "condition",
          "temperature"
        ],
        "type": "object"
      },
...
      "DetailResponse": {
        "properties": {
          "address": {
            "type": "string"
          },
          "current": {
            "$ref": "#/components/schemas/CurrentResponse"
          },
          "farmId": {
            "format": "uuid",
            "type": "string"
          },
          "forecast": {
            "items": {
              "$ref": "#/components/schemas/ForecastResponse"
            },
            "type": "array"
          },
          "observedAt": {
            "format": "date-time",
            "type": "string"
          },
          "partial": {
            "$ref": "#/components/schemas/PartialResponse"
          }
        },
        "required": [
          "address",
          "current",
          "farmId",
          "forecast",
          "observedAt",
          "partial"
        ],
        "type": "object"
      },
...
      "ForecastResponse": {
        "properties": {
          "condition": {
            "$ref": "#/components/schemas/ConditionResponse"
          },
          "date": {
            "format": "date",
            "type": "string"
          },
          "dayOfWeek": {
            "type": "string"
          },
          "maxTemperature": {
            "format": "int32",
            "type": "integer"
          },
          "minTemperature": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "condition",
          "date",
          "dayOfWeek"
        ],
        "type": "object"
      },
...
      "LikeToggleResponse": {
        "properties": {
          "likeCount": {
            "format": "int64",
            "type": "integer"
          },
          "liked": {
            "type": "boolean"
          }
        },
        "required": [
          "likeCount",
          "liked"
        ],
        "type": "object"
      },
...
      "PartialResponse": {
        "properties": {
          "degraded": {
            "type": "boolean"
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "degraded",
          "missing"
        ],
        "type": "object"
      },
...
      "PostDetailResponse": {
        "properties": {
          "author": {
            "$ref": "#/components/schemas/AuthorResponse"
          },
          "body": {
            "type": "string"
          },
          "commentCount": {
            "format": "int64",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "cropId": {
            "format": "uuid",
            "type": "string"
          },
          "cropName": {
            "type": "string"
          },
          "farmingRecordId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "imageUrls": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "likeCount": {
            "format": "int64",
            "type": "integer"
          },
          "likedByMe": {
            "type": "boolean"
          },
          "postType": {
            "enum": [
              "GENERAL",
              "QUESTION"
            ],
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "author",
          "body",
          "commentCount",
          "createdAt",
          "cropId",
          "cropName",
          "id",
          "imageUrls",
          "likeCount",
          "likedByMe",
          "postType",
          "title"
        ],
        "type": "object"
      },
      "PostIdResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
      "PostPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PostSummaryResponse"
            },
            "type": "array"
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "PostSummaryResponse": {
        "properties": {
          "author": {
            "$ref": "#/components/schemas/AuthorResponse"
          },
          "bodyPreview": {
            "type": "string"
          },
          "commentCount": {
            "format": "int64",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "cropId": {
            "format": "uuid",
            "type": "string"
          },
          "cropName": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "likeCount": {
            "format": "int64",
            "type": "integer"
          },
          "likedByMe": {
            "type": "boolean"
          },
          "postType": {
            "enum": [
              "GENERAL",
              "QUESTION"
            ],
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "author",
          "bodyPreview",
          "commentCount",
          "createdAt",
          "cropId",
          "cropName",
          "id",
          "likeCount",
          "likedByMe",
          "postType",
          "title"
        ],
        "type": "object"
      },
...
      "SavePostRequest": {
        "properties": {
          "body": {
            "type": "string"
          },
          "cropId": {
            "format": "uuid",
            "type": "string"
          },
          "farmingRecordId": {
            "format": "uuid",
            "type": "string"
          },
          "mediaIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 5,
            "minItems": 0,
            "type": "array"
          },
          "postType": {
            "enum": [
              "GENERAL",
              "QUESTION"
            ],
            "type": "string"
          },
          "title": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "body",
          "cropId",
          "mediaIds",
          "postType",
          "title"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "access-token": {
        "bearerFormat": "JWT",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "ChamChamCham Backend API Documentation",
    "title": "ChamChamCham API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/community/comments/{commentId}": {
      "delete": {
        "operationId": "deleteComment",
        "parameters": [
          {
            "in": "path",
            "name": "commentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUnit"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "community-controller"
        ]
      }
    },
    "/api/v1/community/posts": {
      "get": {
        "operationId": "listPosts",
        "parameters": [
          {
            "in": "query",
            "name": "authorMemberId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cropId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "postType",
            "required": false,
            "schema": {
              "enum": [
                "GENERAL",
                "QUESTION"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "likedOnly",
            "required": false,
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "mineOnly",
            "required": false,
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "LATEST",
              "enum": [
                "LATEST",
                "LIKE",
                "COMMENT",
                "POPULAR"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePostPageResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "community-controller"
        ]
      },
      "post": {
        "operationId": "createPost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavePostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePostIdResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "community-controller"
        ]
      }
    },
    "/api/v1/community/posts/{postId}": {
      "delete": {
        "operationId": "deletePost",
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUnit"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "community-controller"
        ]
      },
      "get": {
        "operationId": "getPost",
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePostDetailResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "community-controller"
        ]
      },
      "patch": {
        "operationId": "updatePost",
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavePostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePostIdResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "community-controller"
        ]
      }
    },
    "/api/v1/community/posts/{postId}/comments": {
      "get": {
        "operationId": "listComments",
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCommentPageResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "community-controller"
        ]
      },
      "post": {
        "operationId": "createComment",
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCommentIdResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "community-controller"
        ]
      }
    },
    "/api/v1/community/posts/{postId}/like-toggle": {
      "post": {
        "operationId": "toggleLike",
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseLikeToggleResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "community-controller"
        ]
      }
    },
...
    "/api/v1/farming-reports/{reportId}": {
      "get": {
        "operationId": "getDetail",
        "parameters": [
          {
            "in": "path",
            "name": "reportId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDetailResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "farming-cycle-report-controller"
        ]
      }
    },
...
    "/api/v1/farming-reports/{reportId}/work-types/{workType}": {
      "get": {
        "operationId": "getDetail_1",
        "parameters": [
          {
            "in": "path",
            "name": "reportId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workType",
            "required": true,
            "schema": {
              "enum": [
                "PLANTING",
                "WATERING",
                "FERTILIZING",
                "PEST_CONTROL",
                "WEEDING",
                "PRUNING",
                "HARVEST",
                "ETC"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDetailResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "farming-work-report-controller"
        ]
      }
    },
...
  },
  "security": [
    {
      "access-token": []
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ]
}